+++ types/governance.go
@@ -1297,7 +1297,13 @@ type UpdateMarket struct {
 }
 
 func (n UpdateMarket) IntoProto() *vegapb.UpdateMarket {
-	return &vegapb.UpdateMarket{}
+	var changes *vegapb.UpdateMarketConfiguration
+	if n.Changes != nil {
+		changes = n.Changes.IntoProto()
+	}
+	return &vegapb.UpdateMarket{
+		Changes: changes,
+	}
 }
 
 func (n UpdateMarket) DeepClone() *UpdateMarket {
@@ -1336,6 +1342,39 @@ func (n UpdateMarketConfiguration) DeepClone() *UpdateMarketConfiguration {
 		cpy.RiskParameters = n.RiskParameters.DeepClone()
 	}
 	return cpy
+}
+
+func (n UpdateMarketConfiguration) IntoProto() *vegapb.UpdateMarketConfiguration {
+	riskParams := n.RiskParameters.rpIntoProto()
+	md := make([]string, 0, len(n.Metadata))
+	md = append(md, n.Metadata...)
+
+	var instrument *vegapb.UpdateInstrumentConfiguration
+	if n.Instrument != nil {
+		instrument = n.Instrument.IntoProto()
+	}
+	var priceMonitoring *vegapb.PriceMonitoringParameters
+	if n.PriceMonitoringParameters != nil {
+		priceMonitoring = n.PriceMonitoringParameters.IntoProto()
+	}
+	var liquidityMonitoring *vegapb.LiquidityMonitoringParameters
+	if n.LiquidityMonitoringParameters != nil {
+		liquidityMonitoring = n.LiquidityMonitoringParameters.IntoProto()
+	}
+
+	r := &vegapb.UpdateMarketConfiguration{
+		Instrument:                    instrument,
+		Metadata:                      md,
+		PriceMonitoringParameters:     priceMonitoring,
+		LiquidityMonitoringParameters: liquidityMonitoring,
+	}
+	switch rp := riskParams.(type) {
+	case *vegapb.UpdateMarketConfiguration_Simple:
+		r.RiskParameters = rp
+	case *vegapb.UpdateMarketConfiguration_LogNormal:
+		r.RiskParameters = rp
+	}
+	return r
 }
 
 type UpdateInstrumentConfiguration struct {